Fixed read and take, NOT alive warning, printDebug
Read and take functions have been fixed and now work as suppossed.
Included a warning when reading/taking a change that is not alive.
When reading/taking only 1 element the elements is not changed,
be careful when using that data.
WHen reading/taking multiple elements the not alive cache changes
are not included in the vector.

Added printDebug for additional info.One more verbosity level.


Former-commit-id: 53ab397a6df9dd8eeb9ae47a3c3819b4132f215b