Performance modifications.
Object pool class created to be used by the RTPSMessageCreator methods.
RTPSMessageGroup class created to send group of changes as data or gap.
RTPSMessageGroup_t class that contains 3 messages that can be used each time
is needed from different thread since an instance is created for each Writer and for
each event that needs to send groups of messages.


Former-commit-id: b7d21a737bf3b20480f856683fccbbbf520fd235