StatefulWriter: add new messages to history even if full
If a writer's history fills up, then when a new cache change is
published, the oldest cache change is deleted to make room.

Whether this is correct for a single pub/sub pair is debatable; it may
be preferable to suffer the backpressure.

However, when there are multiple subscribers for a single publisher,
it is almost certainly more correct to allow one "good" subscriber
keep getting reliable data if another "bad" subscriber is either not
keeping up or dies.

With this change, if a second subscriber dies, the CPU usage on the
publisher still increases dramatically (at 1000 msgs/sec with a
history depth of 5000, it goes from ~15% to ~75%), but at least the
data keeps flowing. Doing the history management more efficiently and
detecting the loss of the subscriber might be the next step to reign
in the CPU usage.