Add ability to apply the whitelist to input sockets and locators.
The UDP descriptors have been modified to support flags that
allow the interface white list to be applied to output sockets,
input sockets and/or locators.  By default the white list is
only applied to output sockets, which forces it to bind to
allowed interface IPs.

If applied to input sockets, rather than binding to IP address
any (0.0.0.0), the input socket will bind to the multicast IP
if the locator is multicast or to specific allowed interface IPs.

If applied to locators, then the default unicast locator list
will be filtered against the allowed interface IPs.  It is not
applied in the case where the user has specified a default list.