Incorrect gapList.base field in GAP submessage
Given a gap list of sequence numbers 8, 9, 10, 15, 16, 19, 20, 23, 27, 28
The gapList.base field for the gap message was not set correctly causing the first gap sequence to be skipped by teh receiver.

This change correctly set the gapList.base to be +1 from the gapStart value. in our example: gapStart==8, gapList.base==11 gapList.set[15, 16, 19, 20, 23, 27, 28]
Previously, the gap values would be set incorrectly as such: gapStart==8, gapList.base==8 gapList.set[9, 10, 15, 16, 19, 20, 23, 27, 28] causing the gap seqnum 8 to be missed by the receiver.

Also addressed an issue if the last change requested by the remote reader is an invalid change, the local writer would never send a GAP packet (or anything else for that matter). That was caused by the fact that invalid changes (GAP) were ignored by the Acknack processing.

Issue: #182