Fix communication when only localhost is activated.
We were running into a situation where Fast-RTPS would not
pass any data when all network interfaces were down except for
localhost.  It turns out that Fast-RTPS was ignoring localhost
as a valid communications interface in some circumstances.  This
patch fixes things so that Fast-RTPS will now consider localhost
as a communications interface, allowing this to work.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>