VxWorks support
[1] Fix error in asio error code comparison

eProsima code is making asio error code comparisons in an incorrect and
inconsistent manner, where the lhs and rhs types differ in an unintended
way. This results in a compiler error with an ambiguous use of the equality
operator

This change corrects the comparison code to use the value() method in
order to obtain the actual error code value.

[2] Workaround for missing std::put_time in gcc 4.8.4

The original code made use of std::put_time defined under c++11.
GCC 4.8.4, however, does not support this in its implementation of
the c++11 standard.

Sufficiently equivalent functionality has been added to replace
the original use of std::put_time.

[3] Support cross compilation of EProsima