Vendor FindThreads.cmake from CMake 3.14-rc2. [4790] (#438)
* Vendor FindThreads.cmake from CMake 3.14-rc2.

In order to resolve linking issues that appear when running Fast-RTPS
with ASan [1] due to the asan library providing the symbol whose absense
is being used as a test for the need to link pthreads.

This CMake pull request [2] addresses the issue by using a different
function as the heuristic and has been accepted upstream as part of the
3.14-rc2 release.

This commit drops the version of FindThreads.cmake from that release
into the local CMake modules directory where it should take precedence
over the system-provided one.

This file can be removed when the cmake_minimum_required version is 3.14
or newer as the system version should be sufficient.

It would be nice to modify this module to print a warning when it is no
longer needed. I tried to find a means to check the minimum version
rather than the current cmake version but didn't come up with a way to
do so. Issuing a warning a warning on the current cmake version doesn't
make sense because it would nag users on rolling systems to make a
change incompatible with an older stable system.

[1]: https://wiki.gentoo.org/wiki/AddressSanitizer/Problems#pthread_linking_issues
[2]: https://gitlab.kitware.com/cmake/cmake/merge_requests/2629

Fixes eProsima/Fast-RTPS#430

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>

* Fix include path.

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>

* Vendor CheckForPthreads.c from CMake 3.14-rc2