ReaderProxy history reloaded incorrectly missing GAP message
The problem is invalid changes at the end of the stateful writer history don't get added to a newly connected reader proxy cache causing the reader proxy to miss a lot of invalid changed and not generate the proper GAP message.

The fix is to make sure the missing changes are loaded into the reader proxy history so that the GAP message is properly generated.

Reworked the error message on missing changes to eliminate legitimate cases (UNACK'ed changed or duplicated ACK request).
The ERROR will only show if the change was not found and the change was not ACK'ed previously.

Remove meaningless assert statement.

issue #452