The ParticipantImpl::~ParticipantImpl() must be modified to prevent a discovery callback error.
As implementation currently works the Participant object pass as argument to the callbacks is destroyed before this callbacks are cutoff. Callbacks are cutoff in the RTPSDomain::removeRTPSParticipant so the Participant object pass as argument should at least be alive till this call returns. The file src/cpp/participant/ParticipantImpl.cpp should be modified by reordering:

	// delete(mp_participant); // former position

	if(this->mp_rtpsParticipant != nullptr)
	{
		RTPSDomain::removeRTPSParticipant(this->mp_rtpsParticipant);
	}

	delete(mp_participant); // new position