Refs #5624. Fixed bug when adding change in Publisher History.
When durability qos is volatile, the publisher history removes the recently added changes if no subscriber are connected.
The change has to be added to the keys vector first, so the publisher can find the change and it doesn't log any warning.
It doesn't check if the add_change function fails because the only way for that to happen is that the history is full,
but the publisher previously releases one change on the history in order to be sure the change could be added.