Deprecate RTPSMessageGroup_t in favor of using the CDRMessagePool instead
Remove the early allocation of the CDRMessage_t contained in RTPSMessageGroup_t and replace it with the usage of the CDRMessagePool.
The savings are linearly proportional to the number of Statefull readers / writers. Those use to preallocate CDRMessage_t buffers for the ACKNACK and HEARTBEAT packets resulting in a significant amount of RAM used.
Instead, these will now leverage the CDRMessagePool and allocate a buffer from there on demand only, saving a large amount of memory for lots of statefull endpoints.

Issue: #684