A ProxyHashTable class has been introduce to manage ParticipantProxyData reader and writer proxies collections.
This class was deemed necessary because simplifies the use of a memory_pool by attaching it to the collection that
 actually profits from it and a single class can manage any type of pointer.

The foonathan Segregator concept was introduced to prevent the pooled memory from being use in non-node or
collection initialization allocations.

In order to match the number of expected nodes and the block allocation size the implementation overhead is passed as
an argument to the Segregator object.