Refs #6816. Fixed error.
Fixed error with intraprocess communication and VOLATILE reader. When matched, the changes in history is set as
irrelevant, the intraprocess_gap() failes (reader not discovered writer) and the CacheChangeReader_t is not included in
the ReaderProxy. The low_mark is not updated and this causes not sending GAPs when initial heartbeat is sent again. Also
now the initial heartbeat can sends GAPs.