Refs 7835. Lock on create_participant [7836]
Publisher/subscribers should not access the ComainParticipant pointer
until it is fully constructed (i.e., until create_participant returns).

Some of the tests have pubsubs with listeners that access the participant
pointer when receiving a message. Depending on the timing, the receive
resources can be up and running and start receiving before the pointer
is stored internally. When the listener tries to access this pointer,
the test crashes.

Added a mutex lock to sync both parts (participant creation and
listener)