Do not exit the process if the IP detection fails
It is considered bad style to call exit() within a library function, as
this kind of error "handling" cannot be overridden by the calling
program.

As the function getIPs() has a boolean return type already, and the
Windows version actually uses it to convey success or failure, I
adjusted the behavior for the non-Windows version accordingly.

Signed-off-by: Timo Röhling <timo@gaussglocke.de>