Allow fully qualified name in TypeDescriptor
According to the DDS-XTypes spec 7.5.2.4.10, the name of TypeDescriptor
should be the fully qualified name, which is the concatenation of
module names with the name of a type inside of those modules.

This change will allow an object of DynamicTypeBuilder to set a type
name that contains colon in it.

Signed-off-by: Yangbo Long <yangbo.long.mav@gmail.com>