Allow fully qualified name in TypeDescriptor (#1831)
* Allow fully qualified name in TypeDescriptor

According to the DDS-XTypes spec 7.5.2.4.10, the name of TypeDescriptor
should be the fully qualified name, which is the concatenation of
module names with the name of a type inside of those modules.

This change will allow an object of DynamicTypeBuilder to set a type
name that contains colon in it.

Signed-off-by: Yangbo Long <yangbo.long.mav@gmail.com>

* Try to fix linters

Signed-off-by: Yangbo Long <yangbo.long.mav@gmail.com>

* Add some unit tests for type descriptor names

Signed-off-by: Yangbo Long <yangbo.long.mav@gmail.com>

* Fix linters

Signed-off-by: Yangbo Long <yangbo.long.mav@gmail.com>
(cherry picked from commit 61267e378c9b90f0385cf3e1cf219bb8bbc20ed5)