Small revamp in sqlite3.
gcc 11 pointed out that we could sometimes return a stack
pointer from within sqlite3SelectNew().  That could only
happen in Release mode (where asserts are disabled), but
that is still a problem.  I took a look at upstream sqlite3
(3.36), and saw that this had already been fixed.  This is
a backport of just that fix.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>