Fix various gcc11 warnings (#2321)
* Make sure to set the pid to PID_SENTINEL.

This ensures that the pid always has a value, and gets rid
of a gcc 11 warning about pid being possibly uninitialized.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make sure to check the return value of readUInt*.

Since this can fail, the return value (in the last parameter)
may not get set.  Always check the return value, which gets
rid of some warnings with gcc 11.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Small revamp in sqlite3.

gcc 11 pointed out that we could sometimes return a stack
pointer from within sqlite3SelectNew().  That could only
happen in Release mode (where asserts are disabled), but
that is still a problem.  I took a look at upstream sqlite3
(3.36), and saw that this had already been fixed.  This is
a backport of just that fix.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make sure to use std::fill_n to initialize an array of structs

gcc 11 points out that the array of ListenerStatus structures
are not POD types, and so using a simple memset on them could
be dangerous.  Instead, switch to 'std::fill_n' with this,
which necessitated adding some new constructors to the non-POD
types.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Uncrustify

Signed-off-by: Miguel Company <MiguelCompany@eprosima.com>

* Upgrade sqlite to 3.36

Signed-off-by: Miguel Company <MiguelCompany@eprosima.com>

Co-authored-by: Miguel Company <MiguelCompany@eprosima.com>
(cherry picked from commit 18b0ea398bead501ab841f8d082b44ff887e105a)