Fix warning when converting string to Time_t (#2399)
The Time_t nanosec() method requires a uint32_t to be passed in.
But the previous code was using an int32_t for the nanoseconds,
which was causing a warning in some compilers.  Fix this by
using a uint32_t for the nano variable.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit b5f7fbfe0a5d2ff01a1946e7e86783aa331dfa43)