Fix sequence of nested structs.
It turns out that with the current XTypes code, attempting
to create a sequence of nested structs ends up in a segfault.
See https://github.com/eProsima/Fast-DDS/issues/2969 for the
details.

After tracing through, it looks like what is happening is that
there is some confusion on what should be registered as a "complete"
type vs a "minimal" type, according to XTypes 1.3.  In particular,
what is happening is that the nested structure is being created
as a complete type, but then when we go to look it up, we look
it up as both a complete type and a minimal type.  The former
succeeds, and the latter fails, leading to the crash.

This attempted fix just skips trying to look it up as a minimal
type at all.  This is probably not the correct fix, but it is unclear
whether:

1.  We should be "building" the type as both a complete and minimal
one at build time, or
2.  We should only look it up as either a complete or minimal

I'm looking for feedback on which way to go here.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>