Support discovery of TinyXML2 by pkgconfig
Allow to discovery TinyXML2 by pkg-config is a `-config.cmake` file is not available.

Since tinyxml2 is built with meson in Yocto it's configuration gets deployed by `pkg-config` (because meson cannot generate CMake target files). In this case, the current `FindTinyXML2.cmake` will find out the path of `tinyxml2.h` and of `tinyxml2.so` by looking into the `recipe-sysroot` folder, which path in going to get expanded into the `fastrtps-target.cmake` generated when fastdds is built. The end result is that we're going to have a system-wide deployed config mode package that points to a directory that won't (probably) not exist anymore when other dependent targets are built.

Since it's not possible to force meson to produce a `tinyxml2-target.cmake` file, the most reasonable solution is trying to look for the library by pkg-config.

With this patch applied, the generated `fastrtps-target.cmake` will have a reference to the target `PkgConfig::TinyXML2` instead of a full path.