Immediate first discovery
Waiting for the announcement `period` before the *first* announcement is
surprising behavior, especially with longer intervals.

This commit changes the behavior so that the first announcement happens
almost immediately and subsequent announcements are spaced by the
provided `period`. This better matches user expectations of "connect as
early as possible with periodic retry".

Signed-off-by: Robert Plante <RobertDPlante@gmail.com>